Fix memory leak inside of update report handler.
The issue was with the fact that we were not cancelling the delayed calls
therefore they would stay hanging until the stale time was reached. This meant
that when a lot of reporting was happening there would be N hanging delayed
calls while there should only be one.

We were also not deleting the items of the reports dict after we were done with them.

This fixes: https://github.com/TheTorProject/ooni-backend/issues/28