Use shutil.move instead of os.rename.
This is needed because otherwise it will not be possible to have the reports
dir on a different device other than the archive dir.
When using os.rename if the dst is not on the same device as src a
"OSError: [Errno 18] Invalid cross-device link" is thrown.
shutil.move will by default try to use os.rename, but if it fails it will copy
src to dst and then remove src.