Refresh materialized views concurrently
Prior to this materialized view updates lead to the database tables
related to these views will be locked (also for read), which makes the
explorer UI irresponsive during the 1 hour or so that the materialized
views are being updated.
This makes it happen concurrently, which means that the updated view is
written to a temporary table and then renamed in place, making the old
table still accessible while the update is in progress.