fix(report/handlers): accept more semver versions
The current problem we have, specifically, is that since v1.2.0 the
mobile apps emit full semver including also the build.

This requires several parts of our infrastructure to improve the
regexp being used so to validate also this kind of input.

See also: measurement-kit/measurement-kit#1388

An earlier version of this diff was blessed by @hellais on Slack and
since then I just changed comments.