centrifugation: calculate simhash(body) and simhash(text(body))
It'll enable heuristics that rely on page similarity to mine more
blockpages our of existing OONI dataset.

Draft implementation included non-trivial caching scheme that was trying
to avoid unnecessary computation of simhashes for known body_sha256
values. It turned out that the complexity of the code was providing not
that much benefits for the modern state of OONI data.  There are still
~30% of perfectly same pages coming from different probes, but runtime
de-duplication takes as much real time as computing hashes, moreover,
existing code took non-trivial amount of RAM (over 512M) for computation.

Here are approximate relative numbers of wall clock time needed to
ingest a bucket of OONI data:

No simhash           -- 1.0
Old simhash          -- 4.8
New simhash(body)    -- 4.7
Same + simhash(text) -- 6.0

Current bucket ingestion takes 40...80 minutes, so it'll become 4..8
hours. That's still okay as it's single-CPU and can easily utilize SMP
when it'll be needed.

An open problem is re-ingestion of data stored in http_control and
http_request tables as it'll include re-computation of all the simhashes
without any good reason.

So the de-duplication code is dropped and simhashes are stored together
with body_sha256.