Ignore empty reports on reingestion, closes #141
Empty reports may be result of yaml to json conversion: the yaml report
header was there, but no measurements were submitted as a part of the
yaml file. So the header-only report was converted to empty jsonl report.

Alike reports are ignored on ingestion, as they have no vital columns to
fill (e.g. probe_cc).

Reingestion failed during safety check as these reports "exist", but
their metadata could not be properly preserved as there were no metadata
created for them on ingestion.