bouncer: implement /collector and /test-helpers
This allows us to deploy a bouncer with a much simpler API. So, we can
start using such better API in apps. This makes us ready to switch to
use the OONI orchestra rather than legacy services, since the API added
here is the same API of some OONI orchestra endpoints.