Fix bug in marking of confirmed and anomaly flags
The measurement_meta and report_blob tables are only filled with data on
ingestion and re-ingestion, but not on reprocessing. This resulted in
those tables being empty and the queries returning zero rows therefore
leading to the flags not getting computed on reprocessing.

It was necessary to use these tables previously as we were deriving the
probe_cc of the measurement by doing the join and using that to check if
the fingerprint detected was consistent with the measurement.
Since we are now doing this inside of the http_request_fp feeder we
don't need this JOIN there anymore.