Fixes to return slightly more consistent results in the list_measurements endpoint
We apply the OFFSET LIMIT filters to both the query on the batch
pipeline tables (report, measurements) and on the fastpath table and
then again on the UNION of the two.
We also perform ordering on these two queries **before** the UNION so
that the most recent results will appear first from each of them.

The same limit and order is applied to the UNION so that we the results
are sorted. When we get the final result, which is filtered for
duplicates in python in memory, we will only return the first 100
entries (which are the most recent due to the ordering mentioned
before).

If we were not to return only the first 100 elements we would, in cases
in which there is no overlap between fastpath and batch measurement,
have double the number of rows than what the user asked for in the
limit.