fix(core.py): relax riseupvpn scoring logic
This diff relaxes the riseupvpn scoring logic because we have
determined in https://github.com/ooni/probe-cli/pull/1125#issuecomment-1666364558
that we want this experiment to just collect data useful to riseup.

Note that I needed to change test_aggregation.py as a result of these
changes because the number of ok and anomaly results changed.

Closes https://github.com/ooni/backend/issues/745.