Refactor VPN Routes into its own file
Added new directories v1 and v2 to mirror the paths in the API with the
file structure. Added vpn routes to its own vpn module. Created a v1
module because the path in probe_services.py in the monolith code is
`/api/v1/login`, so we keep the same versioning and path definition