Oonimsm fixes (#983)
* Don't specify protocol for other collectors

* Fix rewriting msm uid

The fastpath was rewriting the measurement uid, generating slight
variations in the timestamp that led to 404 errors when trying to
retrieve the measurement from the explorer

This error was introduced when the feature for writing measurements to
the disk from the fastpath was added.

This commit should not affect the monolith

* Fix bad hour calculation in write measurement to disk

* Simplify hour parsing

* Downgrade to python 3.12 to prevent clickh error

There seems to be an issue with clickhouse when running as a docker
container where it will crash if running in an user other than root.

The problem is that in 3.13 the python standard library they changed
the errors risen by the get user function

see: https://github.com/mymarilyn/clickhouse-driver/pull/462